Quadratsumme

+++ Schleifen/src/Quadratsumme.java
@@ -6,13 +6,13 @@ public class Quadratsumme {
         System.out.print("Geben sie die Endzahl ein: ");
         int ende  = IO.readInt();
 
-        int qSumme;
+        double qSumme;
         int i;
 
 
         for (i = 1; i <= ende; i++) {
 
-           qSumme = (int) Math.pow(i,2);
+           qSumme =  Math.pow(i,2);
             System.out.println(i+ " hoch 2 = "+qSumme);
 
         }
